[chore #161009928] create a default admin on the initial run of the application
