+++ requirements.txt
@@ -18,6 +18,7 @@ itsdangerous==0.24
 Jinja2==2.10
 MarkupSafe==1.0
 more-itertools==4.3.0
+passlib==1.7.1
 pathlib2==2.3.2
 pep8==1.7.1
 pg8000==1.12.3
